app/requirements.txt
aiohttp==3.7.4
aiolimiter==1.0.0
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
async-timeout==3.0.1
attrs==22.2.0
Authlib==1.2.0
bcrypt==4.0.1
Brotli==1.0.9
brotli-asgi==1.3.0
cachetools==5.3.1
cairocffi==1.6.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
code128==0.3
cryptography==40.0.1
cssselect2==0.7.0
defusedxml==0.7.1
discord.py==1.7.3
dnspython==2.3.0
et-xmlfile==1.1.0
fonttools==4.42.1
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.21.1
MarkupSafe==2.1.2
motor==3.1.1
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.1
Pillow==10.0.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pypdf==3.17.1
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-Levenshtein==0.21.1
python-multipart==0.0.5
pytz==2023.3
PyYAML==6.0
pyzbar==0.1.9
rapidfuzz==3.3.0
reportlab==4.0.4
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.23.1
starlette-discord==0.2.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
user-agent==0.1.10
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
